requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-docx
pdfplumber
faiss-cpu
sentence-transformers
numpy
python-dotenv
aiofiles
groq